requirements.txt
aiohttp
black
coverage
flake8
mock
pyflakes
pylint
pytest
pytest-asyncio
pytest-cov
python-dotenv
requests
requests-toolbelt
